feature: Support basic auth for Elasticsearch client
- Changed create statement for cassandra to meet the application's default configurations
- Changed data type for bucketSize
- Explicit type to prevent possible overflow for numbers when calculating QueryOffsetRanges

+++ graphene-reader/src/main/kotlin/com/graphene/reader/store/data/DataFetchHandlerConfig.kt
@@ -60,7 +60,7 @@ data class DataFetchHandlerProperty(
   var columnFamily: String,
   var rollup: Int = 60,
   var maxPoints: Int = Int.MAX_VALUE,
-  var bucketSize: Int,
+  var bucketSize: Short,
   var property: CassandraDataHandlerProperty
 )
