+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/StoreHandlerFactory.kt
@@ -129,7 +129,7 @@ data class DataStoreHandlerProperty(
   var rollup: Int = 60,
   var keyspace: String,
   var columnFamily: String,
-  var bucketSize: Int,
+  var bucketSize: Short,
   var property: CassandraDataHandlerProperty
 )
